requirements.txt
requests==2.32.3
boto3==1.35.99
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.3.4
